refactor ExtensionLoader.init
